move authController to auth directory
